Initial abstraction of justify to own component
