+++ packages/block-editor/src/components/justify-toolbar/README.md
@@ -1,0 +1,3 @@
+## Justify Toolbar
+
+The `JustifyToolbar` component renders a toolbar that displays justify options for the selected block.
